chore(pong): sort game.py into separate files and remove unused imports
