+++ backend/pong/game/ball.py
@@ -1,53 +1,7 @@
-from asgiref.sync import async_to_sync
-from channels.layers import get_channel_layer
-from threading import Lock
-from .serializers import ObjectStateSerializer
+from .data import Vector2, ObjectState
+from .paddle import Paddle
 import copy
-import logging
 import math
-
-class Vector2():
-    def __init__(self, x=0, y=0):
-        self.x = x
-        self.y = y
-
-class ObjectState():
-    def __init__(self, name):
-        self.name = name
-        self.states = []
-
-    # alpha is used for interpolation
-    # possible values are 0.0 - 1.0
-    def append(self, pos, alpha):
-        self.states.append({
-            'pos': copy.copy(pos),
-            'alpha': alpha,
-        })
-
-class Paddle():
-    size = Vector2(7, 45)
-    speed = 250
-
-    def __init__(self, x, y, player_num):
-        self.pos = Vector2(x, y)
-        self.player_num = player_num
-
-    def tick(self, game_info, dt):
-        states = ObjectState('paddle')
-        states.append(self.pos, 0.0)
-
-        player_input = game_info.player_inputs[self.player_num - 1]
-        if player_input.get_input('up'):
-            self.pos.y -= Paddle.speed * dt
-
-        if player_input.get_input('down'):
-            self.pos.y += Paddle.speed * dt
-
-        # clamp pos
-        self.pos.y = min(max(self.pos.y, 0), game_info.game_size.y - Paddle.size.y)
-
-        states.append(self.pos, 1.0)
-        return states
 
 class Ball():
     size = Vector2(7, 7)
@@ -183,56 +137,3 @@ class BallTimer():
         game_info.objects.append(new_ball)
         game_info.objects.remove(self)
         return None
-
-class PlayerInput():
-    def __init__(self):
-        # data race in my code? it's more likely than you think
-        self.input_lock = Lock()
-        self.inputs = {
-            'up': False,
-            'down': False,
-        }
-
-    def set_input(self, input_type, value):
-        self.input_lock.acquire()
-        if input_type in self.inputs and isinstance(value, type(self.inputs[input_type])):
-            self.inputs[input_type] = value
-
-        else:
-            print(f'unknown input {input_type} and value {value}')
-
-        self.input_lock.release()
-
-    def get_input(self, input_type):
-        self.input_lock.acquire()
-        value = self.inputs[input_type]
-        self.input_lock.release()
-
-        return value
-
-class GameLogic():
-    sec_per_frame = 1 / 60
-    ms_per_frame = sec_per_frame * 1000
-
-    def __init__(self):
-        self.ended = False
-        self.player_inputs = [PlayerInput(), PlayerInput()]
-        self.game_size = Vector2(400, 240)
-        self.score = [0, 0]
-        self.win_score = 5
-        self.objects = [
-            Paddle(25, 0, 1), # left paddle
-            Paddle(self.game_size.x - Paddle.size.x - 25, 0, 2), # right paddle
-            Ball(self.game_size.x / 2, self.game_size.y / 2, 1, math.sin(math.radians(45)), 200),
-        ]
-
-    # returns an array of objects for the client to render
-    def tick(self, dt):
-        states = []
-        objlist = self.objects.copy() # this is done to prevent processing new objects created in the same tick
-        for obj in objlist:
-            obj_state = obj.tick(self, dt)
-            if obj_state != None:
-                states.append(ObjectStateSerializer(obj_state).data)
-
-        return states
